# Enable CodeGenHashStability tests by default on all platforms