--- a/tools/clang/unittests/HLSL/CompilerTest.cpp
+++ b/tools/clang/unittests/HLSL/CompilerTest.cpp
@@ -4271,12 +4271,7 @@ TEST_F(CompilerTest, DISABLED_ManualFileCheckTest) {
   }
 }
 
-
-#ifdef _WIN32
 TEST_F(CompilerTest, CodeGenHashStability) {
-#else
-TEST_F(CompilerTest, DISABLED_CodeGenHashStability) {
-#endif
   CodeGenTestCheckBatchHash(L"");
 }