On my local machine the CodeGenHashStability test takes 15 seconds to run. While that is longer than the entire rest of the test suite, it makes the full test suite run only take 17 seconds, which is still reasonable.

This change then goes a step further and breaks up the CodeGenHashStability test into multiple smaller tests. With that refactoring the full test suite drops from 17s to 10s. The end result is that enabling CodeGenHashStability costs only 1s on Release+Aserts builds.

Here is a rough breakdown of the performance impact of this change on test time for the full test suite:

```
┌───────────────┬───────────────┬───────────────┐
│               │Release+Asserts│ Debug+Asserts │
├───────────────┼───────────────┼───────────────┤
│ Before        │      9s       │      57s      │
├───────────────┼───────────────┼───────────────┤
│ After         │     10s       │      88s      │
└───────────────┴───────────────┴───────────────┘
```

We should enable this test case by default and evaluate ways to better parallelize this test so that so that we can run it more efficiently.