@llvm-beanz 2023-08-16
> Why remove the priority property? Otherwise LGTM

The priority property prevents the test from running in the default case. This change’s intent is to make the test run with all the other tests (default priority 0), and to split it into multiple test cases that can run in parallel so that the impact on overall testing time is reduced.